+++ src/localizationAndCorrespondance.cpp
@@ -5,7 +5,7 @@
 
 namespace get3D2DCorrespondance{
 
-    correspondance get_correspondace(std::vector<cv::DMatch>& filtered_matches, std::vector<cv::KeyPoint>& kp_current, std::vector<cv::DMatch>& prev_2D_filtered_matches, std::vector<cv::Point3d>& prev_3D_points){
+    correspondance get_correspondace(std::vector<cv::DMatch>& filtered_matches, std::vector<cv::KeyPoint>& kp_current, std::vector<cv::DMatch>& prev_2D_filtered_matches, std::vector<cv::Point3f>& prev_3D_points){
 
         std::unordered_map<int, int> trainIdx_to_3d_idx;
         correspondance result_3d_2d;
@@ -15,7 +15,7 @@ namespace get3D2DCorrespondance{
             const auto& m = prev_2D_filtered_matches[idx];
             trainIdx_to_3d_idx[m.trainIdx] = idx;
         }
-        std::vector<cv::Point3d> common_3d_pts;
+        std::vector<cv::Point3f> common_3d_pts;
         std::vector<cv::Point2f> corresponding_2d_pts;
 
         for (const auto& m : filtered_matches)
@@ -50,14 +50,13 @@ namespace get3D2DCorrespondance{
 
 
 namespace computeTransformation{
-    resultTransformation computePnP(std::vector<cv::Point3d>& obj_points, std::vector<cv::Point2f>& img_points, cv::Mat& K, cv::Mat& dist_coeffs){
+    resultTransformation computePnP(std::vector<cv::Point3f>& obj_points, std::vector<cv::Point2f>& img_points, cv::Mat& K, cv::Mat& dist_coeffs){
         resultTransformation solvedTransformations;
         std::vector<int> inliers;
         cv::Mat rotation_vector;
         bool result = cv::solvePnPRansac(obj_points, img_points, K, dist_coeffs, rotation_vector, solvedTransformations.translation_vector,false, 100, 8.0, 0.99, inliers, cv::SOLVEPNP_ITERATIVE);
         
         cv::Rodrigues(rotation_vector, solvedTransformations.rotation_matrix);
-
         return solvedTransformations;
     }
 }
